fix(request): require qs
